+++ tests/test_tile.py
@@ -7,6 +7,10 @@ class TestTiles(unittest.TestCase):
         self.assertEqual(tile.letter, 'A')
         self.assertEqual(tile.value, 1)
 
+    def test_repr(self):
+        tile = Tile('A', 1)
+        self.assertEqual(repr(tile), 'A:1')
+
 if __name__ == '__main__':
 
     unittest.main()